% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_same.R
\name{all_same}
\alias{all_same}
\title{Check if all elements are unique}
\usage{
all_same(x)
}
\arguments{
\item{x}{Numeric or character vector}
}
\value{
logical indicating T for unique, F for non-unique (vector is all the same value)
}
\description{
Check if all elements are unique
}
\examples{
all_unique(c(1,2,3,4))
all_unique(c(1,1,1,1))
}
